refactor: 예매 내역 관련 클래스들을 별도의 파일로 분리
